fixing bug for not found

+++ apps/projects/views.py
@@ -1,5 +1,5 @@
 from django.shortcuts import render, redirect
-
+from django.http import HttpResponse
 from apps.projects.models import Project
 from .forms import ProjectModelForm
 from apps.core import utils
@@ -14,6 +14,8 @@ def projects(request):
 
 def single_project(request, uuid):
     project = utils.get_object_or_404(model=Project, uuid=uuid)
+    if project is False:
+        return HttpResponse("Not Found")
     tags = project.tags.all()
     return render(
         request,
@@ -38,6 +40,10 @@ def create_project(request):
 
 def delete_project(request, uuid):
     project = utils.get_object_or_404(model=Project, uuid=uuid)
+
+    if project is False:
+        return HttpResponse("Not Found")
+
     if request.method == "POST":
         project.delete()
         return redirect("projects:projects")
@@ -46,6 +52,10 @@ def delete_project(request, uuid):
 
 def update_project(request, uuid):
     project = utils.get_object_or_404(model=Project, uuid=uuid)
+
+    if project is False:
+        return HttpResponse("Not Found")
+
     form = ProjectModelForm(instance=project)
     if request.method == "POST":
         form = ProjectModelForm(data=request.POST, instance=project)
